+++ src/main/java/com/blog/mappers/TagMapper.java
@@ -17,11 +17,11 @@ public interface TagMapper {
   TagDto toTagResponse(Tag tag);
 
   @Named("calculatePostCount")
-  default Integer calculatePostCount(Set<Post> posts) {
+  default long calculatePostCount(Set<Post> posts) {
     if (posts == null) {
       return 0;
     }
-    return (int) posts.stream().filter(post -> PostStatus.PUBLISHED.equals(post.getPostStatus()))
+    return posts.stream().filter(post -> PostStatus.PUBLISHED.equals(post.getPostStatus()))
         .count();
   }
 }
